QMM011 F + B screen

+++ nts.uk/uk.pr/pr.ctx/pr.core/nts.uk.ctx.pr.core.app/src/main/java/nts/uk/ctx/pr/core/app/command/laborinsurance/EmpInsurBusBurRatioCommand.java
@@ -1,7 +1,5 @@
 package nts.uk.ctx.pr.core.app.command.laborinsurance;
 
-import java.math.BigDecimal;
-
 import lombok.Value;
 
 @Value
@@ -9,8 +7,8 @@ public class EmpInsurBusBurRatioCommand {
     
     private String hisId;
     private int empPreRateId;
-    private BigDecimal indBdRatio;
-    private BigDecimal empContrRatio;
+    private Long indBdRatio;
+    private Long empContrRatio;
     private int perFracClass;
     private int busiOwFracClass;
